Started rework of connection
So we need to move all the calls to a seperate single file which is "client.js" that will be the file that will handle all the data request and everything else. The reason for this is so that we keep the socket connection during the whole game, otherwise when we change scene it will create a new connection. We need to be able to get the user data (like name, friends rank etc from the index file somehow)